+++ NEWS.md
@@ -1,5 +1,7 @@
-# attachment 0.2.2
+# attachment 0.2.2.9000
 
+* Add NAMESPACE if missing with `att_amend_desc(document = TRUE)`
+* Add DESCRIPTION with empty skeleton if missing with `att_amend_desc()`
 * Default to remove NAMESPACE before updating it in `att_from_namespace()`
 * Fix detection of multiple render outputs in Rmd with `att_from_rmd()`
 * Allow vector of R files in `att_from_rscripts()`